Modify kdcnn class (neurodata/kdg#4)
This amended class was used to experiment on the all FC layaers vs. penultimate layer and polytope member threshold T. In this class, the CNN is trained outside and the trained model is taken as an input to the class. This enables to use the same trainedCNN for all the experiments involving kdcnn.